requirements.txt
python-dotenv>=0.19.0
openai>=1.0.0
tavily-python>=0.1.0
aiohttp>=3.8.0
asyncio>=3.4.3 